feat: webpack.config update to ts version
- webpack配置文件迁移至ts版本
