+++ package.json
@@ -7,8 +7,10 @@
     "@babel/core": "^7.16.0",
     "@babel/preset-env": "^7.16.0",
     "@babel/preset-react": "^7.16.0",
+    "@types/node": "^16.11.6",
     "@types/react": "^17.0.33",
     "@types/react-dom": "^17.0.10",
+    "@types/webpack": "^5.28.0",
     "babel-loader": "^8.2.3",
     "clean-webpack-plugin": "^4.0.0",
     "css-loader": "^6.5.0",
@@ -21,6 +23,7 @@
     "sass-loader": "^12.3.0",
     "style-loader": "^3.3.1",
     "ts-loader": "^9.2.6",
+    "ts-node": "^10.4.0",
     "typescript": "^4.4.4",
     "url-loader": "^4.1.1",
     "webpack": "^5.61.0",
@@ -28,8 +31,8 @@
     "webpack-dev-server": "^4.4.0"
   },
   "scripts": {
-    "dev": "webpack serve",
-    "build": "webpack"
+    "dev": "webpack serve --config ./webpack.config/webpack.config.ts",
+    "build": "webpack --config ./webpack.config/webpack.config.ts"
   },
   "dependencies": {
     "react": "^17.0.2",
