+++ tsconfig.json
@@ -1,13 +1,14 @@
 {
   "compilerOptions": {
     "target": "es5",                                     /* Set the JavaScript language version for emitted JavaScript and include compatible library declarations. */
-    "jsx": "react",                                /* Specify what JSX code is generated. */
-    "experimentalDecorators": true,                   /* Enable experimental support for TC39 stage 2 draft decorators. */
+    "jsx": "react",                                      /* Specify what JSX code is generated. */
+    "experimentalDecorators": true,                      /* Enable experimental support for TC39 stage 2 draft decorators. */
     "module": "commonjs",                                /* Specify what module code is generated. */
     "esModuleInterop": true,                             /* Emit additional JavaScript to ease support for importing CommonJS modules. This enables `allowSyntheticDefaultImports` for type compatibility. */
+    "allowSyntheticDefaultImports": true,
     "forceConsistentCasingInFileNames": true,            /* Ensure that casing is correct in imports. */
     "strict": true,                                      /* Enable all strict type-checking options. */
-    "skipLibCheck": true,                                 /* Skip type checking all .d.ts files. */
+    "skipLibCheck": true,                                /* Skip type checking all .d.ts files. */
     "baseUrl": "./",
     "paths": {
       "@/*": ["./src/*"]
