+++ webpack.config/loaders/img-loader.ts
@@ -1,6 +1,7 @@
+
 const loader = {
   test: /\.(png|jpe?g|gif)$/i,
   use: ['file-loader']
 };
 
-module.exports = loader;
+export default loader;
